fix(topology): x-message-ttl now passed correclty (previous we can pass float instead int; rabbitmq not allow float)
https://www.rabbitmq.com/docs/ttl#per-queue-message-ttl

+++ dramatiq_kombu_broker/topology.py
@@ -70,7 +70,7 @@ class DefaultDramatiqTopology:
 
         # The value of the TTL argument or policy must be a non-negative integer (equal to or greater than zero),
         # describing the TTL period in milliseconds.
-        ttl = self.dead_letter_message_ttl.total_seconds() * 1000
+        ttl = int(self.dead_letter_message_ttl.total_seconds() * 1000)
         return {
             "x-message-ttl": ttl,
         }
